requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

numpy==1.24.4
Flask
gunicorn
easyocr
spacy
pillow
opencv-python-headless
openpyxl
nameparser
torch==2.0.0+cpu
torchvision==0.15.1+cpu
